Improving spatial filter
Excluding the 4 diagonal pixels creates some artifacts that are easily
avoidable by just allowing the polar filter to work like a normal polar
filter.

This commit adds these 4 pixels back into the equation while keeping the
12-tap regression intact since the 16-tap regression is almost always
worse.

The code is admittedly a bit ugly but it is what it is.