Improving mix_coeff
Previously, the regressions could only contribute up to 50% of the final
output, and the correlation was squared. I'm changing this to 80% max
with corr^8. This change allows the regressions to contribute more on
regions where we're confident the output will be good, and it also makes
them contribute less when they would produce subpar results.