Increasing mixing coefficient to 1.0
This is kind of a risky move and I may have to revert this later, but
it provides higher SSIM at the expense of all other metrics.

The bet here is just that if someone is willing to use a chroma shader,
he/she probably wants the image to look as tight as possible, and
allowing the regressions to contribute the entire pixel at
max correlation does exactly that.

The other metrics probably drop because doing a simple linear regression
with the 12 nearest pixels is obviously far from ideal, but it is what
it is.