AR changes, ringing good
Seriously though: some overshooting is necessary on HF regions, I was
under the impression that the regression could deal with it on its own
but I was wrong. After comparing against the "old cfl" the "new cfl" was
lacking some fine-detail sometimes.

This will obviously make actual ringing more easily visible but it is
what it is.