Simplifying downscaling logic again
This offers a small performance increase for all scaling factors while
providing the exact same quality at 2x (normal 4:2:0 use-case).

At 4x (when using a luma doubler) it is a bit worse but nothing
noticeable on real content. A nice side-effect is that the problem with
vanishing line-art seems to be a bit less severe now (I think it's due
to lower correlation, the older downscaling code was blurrier which
made the correlation between planes higher).

Again I may or may not revert this at some point, but I'm reasonably
happy with the change and it also reduces bloat a bit.