Adding experiments
CfL_Prediction_dev is the same as the current lite variant but with the spatial
filter on its own shader pass. This allows for it to be corrected independently
from the main regression pass. In theory this should be "more correct".

CfL_Prediction_dev_no_spatial_filter removes the spatial filter completely and relies
on mpv to upscale chroma for it. Use it with something like --cscale=ewa_robidoux
--cscale-param1=0.0 --cscale-param2=0.75 --cscale-antiring=0.8 to get an output
that's as sharp as the original shader.