Promoting dev, and simplifying lite further.
Normal version has been updated to use Hermite to downscale. This was
necessary to stop the downscaling pass from introducing too much aliasing
on non-centre chromaloc sources. The previous arrangement was completely
broken on anything other than centre. This makes the shader worse than
before when you're watching a source with bilinear/box chroma, but
hopefully that's not common enough in video (it is pretty common for
images though).

Likewise, the lite variant has been reverted to how it was before it was
capable of reacting to chromaloc changes. This was done to prevent it
from creating pure bogus results. The downscaling step matches box at
0.5x.

This will likely make the lite variant better than the main variant at
centre chromaloc images, but that's a compromise I'm willing to live with.

PS: There's a huge number of videos out there that are supposed to have
left chromaloc but are actually centre. There's nothing we can do about
this and in these cases the lite variant will do better.

PS2: Many thanks to Kasper for trying to figure this shit out with me.